requirements.txt
azure-functions==1.24.0
WeasyPrint==66.0
